+++ src/auth/UserAuth/AuthUser.jsx
@@ -1,39 +1,29 @@
 // import axios from "axios";
 import { createAsyncThunk } from '@reduxjs/toolkit';
 import axios from 'axios';
-axios.defaults.baseURL = 'https://petly-backend-23cb.onrender.com/api';
+// axios.defaults.baseURL = 'https://petly-backend-23cb.onrender.com/api';
 
-const token = {
-    set(token) {
-        axios.defaults.headers.common.Authorization = `Bearer ${token}`;
-    },
-    unset() {
-        axios.defaults.headers.common.Authorization = '';
-    },
-};
+// const token = {
+//     set(token) {
+//         axios.defaults.headers.common.Authorization = `Bearer ${token}`;
+//     },
+//     unset() {
+//         axios.defaults.headers.common.Authorization = '';
+//     },
+// };
 
-export const userCurrent = createAsyncThunk(
-    '/user',
+// export const fetchUser = createAsyncThunk(
+//     'user',
 
-    async (_, { rejectWithValue, getState }) => {
-        const { user } = getState();
-        const userToken = user.token;
-        // console.log(userToken);
-
-        token.set(userToken);
-
-        if (userToken === null) {
-            return rejectWithValue();
-        }
-
-        try {
-            const { data } = await axios.get('/user');
-            return data;
-        } catch (error) {
-            return rejectWithValue(error.message);
-        }
-    }
-);
+//     async (_, { rejectWithValue }) => {
+//         try {
+//             const { data } = await axios.get('/user');
+//             return data;
+//         } catch (error) {
+//             return rejectWithValue(error.message);
+//         }
+//     }
+// );
 
 export const loggedOut = createAsyncThunk(
     'auth/logout',
@@ -41,7 +31,7 @@ export const loggedOut = createAsyncThunk(
         try {
             const { data } = await axios.post('/auth/logout');
             // console.log(data);
-            token.unset();
+            // token.unset();
             return data;
         } catch (error) {
             return thunkApi.rejectWithValue(error.message);
@@ -49,22 +39,22 @@ export const loggedOut = createAsyncThunk(
     }
 );
 
-export const updateUser = createAsyncThunk(
-    'user/update',
-    async (userId, { rejectWithValue }) => {
-        console.log(userId);
-        const result = {
-            email: userId.newEmail,
-            name: userId.newName,
-            birthday: userId.newBirthday,
-            city: userId.nweCity,
-            phone: userId.newPhone,
-        };
-        try {
-            const { data } = await axios.patch(`user/${userId}`, result);
-            return data;
-        } catch (error) {
-            return rejectWithValue(error.message);
-        }
-    }
-);
+// export const updateUser = createAsyncThunk(
+//     'user/update',
+//     async (userId, { rejectWithValue }) => {
+//         console.log(userId);
+//         const result = {
+//             email: userId.newEmail,
+//             name: userId.newName,
+//             birthday: userId.newBirthday,
+//             city: userId.nweCity,
+//             phone: userId.newPhone,
+//         };
+//         try {
+//             const { data } = await axios.patch(`user/${userId}`, result);
+//             return data;
+//         } catch (error) {
+//             return rejectWithValue(error.message);
+//         }
+//     }
+// );
