+++ src/pages/User/UserPage.jsx
@@ -1,24 +1,33 @@
-// import { useEffect } from 'react';
-import { useDispatch } from 'react-redux';
-import { getUserOperation } from 'redux/User/userOperations';
+import { useEffect } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
 
-import UserData from 'components/UserData/UserData';
+// import { getUserOperation } from 'redux/User/userOperations';
+import { fetchUser } from '../../redux/User/userOperations';
+
+import UserData from '../../components/UserData/UserData';
 import PetsData from 'components/PetsData/PetsData';
 import { Container } from './UserPage.styled';
+import { getLoader } from 'redux/User/selectors';
+import { Loader } from 'components/Utils/Loader/Loader';
 
 const UserPage = () => {
     const dispatch = useDispatch();
+    const isLoading = useSelector(getLoader);
 
-    dispatch(getUserOperation());
+    useEffect(() => {
+        dispatch(fetchUser());
+    }, [dispatch]);
 
-    // useEffect(() => {
-    //     console.log('useEffect start us');
-    //     dispatch(getUserOperation());
-    // }, [dispatch]);
     return (
         <Container>
-            <UserData />
-            <PetsData />
+            {isLoading ? (
+                <Loader />
+            ) : (
+                <>
+                    <UserData />
+                    <PetsData />
+                </>
+            )}
         </Container>
     );
 };
