+++ src/redux/User/selectors.js
@@ -1,2 +1,3 @@
 export const getUser = ({ user }) => user.user; 
+console.log(getUser)
 export const getLoader = ({ user }) => user.isLoading;
